refactored and updated configs
