+++ jbrowse/config/dev.config.js
@@ -4,14 +4,14 @@
  * Licensed under the Apache License, Version 2.0: http://www.apache.org/licenses/LICENSE-2.0
  */
 
-const watchConfig = require('./node_modules/@labkey/build/webpack/watch.config')
+const devConfig = require('../node_modules/@labkey/build/webpack/dev.config')
 
-const entryPoints = require('./src/client/entryPoints');
-const constants = require('./node_modules/@labkey/build/webpack/constants');
+const entryPoints = require('../src/client/entryPoints');
+const constants = require('../node_modules/@labkey/build/webpack/constants');
 const webpack = require('webpack');
 const NodePolyfillPlugin = require("node-polyfill-webpack-plugin")
 
-const clientConfig = watchConfig
+const clientConfig = devConfig
 
 clientConfig.resolve.fallback =
 {
