+++ jbrowse/config/prod.config.js
@@ -4,14 +4,14 @@
  * Licensed under the Apache License, Version 2.0: http://www.apache.org/licenses/LICENSE-2.0
  */
 
-const devConfig = require('./node_modules/@labkey/build/webpack/dev.config')
+const prodConfig = require('../node_modules/@labkey/build/webpack/prod.config')
 
-const entryPoints = require('./src/client/entryPoints');
-const constants = require('./node_modules/@labkey/build/webpack/constants');
+const entryPoints = require('../src/client/entryPoints');
+const constants = require('../node_modules/@labkey/build/webpack/constants');
 const webpack = require('webpack');
 const NodePolyfillPlugin = require("node-polyfill-webpack-plugin")
 
-const clientConfig = devConfig
+const clientConfig = prodConfig
 
 clientConfig.resolve.fallback =
 {
